Deleted GameSettings Class. Added new GameElements. GameElement updated.
The class GameSettings wasn't used and therefore was deleted. New
GameElements such as: HealingPotion, Thug and Treasure were added to the
Game so that it could load all the rooms. The generateMap() function in
the Room class was changed with the objective of distinguish doors with
Keys and doors with no Keys attached. The abstract function isPickable()
was set to TRUE in all pickable objects.

MAJOR bug #2 SOLVED: The game is now rendering all rooms from any files
without erros or warnings.

------------------------------- DETAILS ------------------------------

-> Deleted GameSettings Class.
-> Added HealingPotion.
-> Added Thug.
-> Added Treasure.
-> All pickable object's isPickable() set to true.
-> Fabric Method in GameElement.class updated to include new
GameElements.
-> Changed generateMap() function, adding the possibility for the
GameEngine to distinguish a Door with no key attached and a Door with a
key attached from the file.